backend/scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.4",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.0",
    "tqdm>=4.67.1",
]
